+++ integrations/gitlab-v2/CHANGELOG.md
@@ -7,12 +7,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- towncrier release notes start -->
 
-## 0.1.7-dev (2025-04-14)
+## 0.1.10-dev (2025-04-14)
 
 
 ### Features
 
 - Add support for gitlab members
+
+
+## 0.1.8-dev (2025-04-14)
+
+
+### Features
+
+- Added support for live events for folder kind
+
+
+## 0.1.7-dev (2025-04-14)
+
+
+### Improvements
+
+- Added support for resolving `file://` references in parsed JSON and YAML files.
 
 
 ## 0.1.6-dev (2025-04-09)